+++ plugins/gui/src/file_manager/import_netlist_dialog.cpp
@@ -1,6 +1,7 @@
 #include "gui/file_manager/import_netlist_dialog.h"
 
 #include "gui/gui_utils/graphics.h"
+#include "gui/export/import_project_dialog.h"
 #include "gui/gatelibrary_management/gatelibrary_selection.h"
 
 #include <QCheckBox>
@@ -14,7 +15,6 @@
 #include <QLabel>
 #include <QLineEdit>
 #include <QPushButton>
-#include <QRegularExpression>
 #include <QSpacerItem>
 #include <QStyle>
 
@@ -50,7 +50,7 @@ namespace hal
         layout->setRowStretch(irow - 1, 20);
 
         layout->addWidget(new QLabel("Location of project directory:", this), irow++, 0, Qt::AlignLeft);
-        setSuggestedProjectDir(filename);
+        mProjectdir = ImportProjectDialog::suggestedProjectDir(filename);
 
         QFrame* frameProjectdir    = new QFrame(this);
         QHBoxLayout* layProjectdir = new QHBoxLayout(frameProjectdir);
@@ -80,22 +80,6 @@ namespace hal
         connect(dbb, &QDialogButtonBox::accepted, this, &QDialog::accept);
         connect(dbb, &QDialogButtonBox::rejected, this, &QDialog::reject);
         layout->addWidget(dbb, irow++, 0, Qt::AlignRight);
-    }
-
-    void ImportNetlistDialog::setSuggestedProjectDir(const QString& filename)
-    {
-        mProjectdir = filename;
-        mProjectdir.remove(QRegularExpression("\\.\\w*$"));
-
-        QString basedir = mProjectdir;
-        int count       = 2;
-
-        for (;;)    // loop until non existing directory found
-        {
-            if (!QFileInfo(mProjectdir).exists())
-                return;
-            mProjectdir = QString("%1_%2").arg(basedir).arg(count++);
-        }
     }
 
     QString ImportNetlistDialog::projectDirectory() const